requirements.txt
aenum==3.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
defusedxml==0.7.1
dumpit==0.6.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
Marine-Traffic-API==0.20.1
MarkupSafe==1.1.1
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
termcolor==1.1.0
terminaltables==3.1.0
twilio==6.55.0
ujson==4.0.2
urllib3==1.26.4
Werkzeug==1.0.1
